requirements.txt
# Core Framework
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Database (Oracle EBS R12)
oracledb==1.4.1

# Configuration
python-dotenv==1.0.0

# Data Validation
pydantic==2.4.2

# HTTP Client (Ollama integration)
requests==2.31.0

# Logging & Observability
python-json-logger==2.0.7

# ML Intent Classification
scikit-learn==1.3.2
numpy==1.24.3

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pylint==3.0.2
